refactor: parameter type

+++ server1/src/main/java/sinhan/server1/domain/user/controller/UserController.java
@@ -79,7 +79,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{family-sn}")
-    public ApiUtils.ApiResult disconnectFamily(@PathVariable int familySn) throws Exception {
+    public ApiUtils.ApiResult disconnectFamily(@PathVariable long familySn) throws Exception {
         UserInfoResponse userInfo = jwtService.getUserInfo(jwtService.getAccessToken());
 
         if (!isFamilyUser(familySn)) {
